Reorganise application action tests
There’s a lot of these now, so it makes sense to split them up
